restructure args
