+++ Source/Extensions/Blazorise.Maps/Enums/MapProvider.cs
